net: ldpaa_eth: connect to multiple PHYs/retimers
There are cases where multiple retimers are needed for a single port.
This is the case of the M8 mezzanine card used on the LX2160AQDS boards,
where 2 Inphi IN112525 S05 retimers are needed for a  100G port.

In this case we need to specify 2 PHYs in the phy-handle property so
that both retimers are initialized.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>