LFU-182-1 arm: imx8ulp: Probe the S400 MU after u-boot relocation
The S400 MU is only probed before u-boot relocation but not after
the relocation. This will cause problem when communication with S400
since the variables before relocation are used.

To fix the issue, probe the S400 MU in arch_early_init_r and select
the CONFIG_ARCH_EARLY_INIT_R by default

Signed-off-by: Ye Li <ye.li@nxp.com>