LFU-193-2 usb: xhci-imx8m: Update iMX8MP XHCI driver to support new DTS node
We will sync the USB DTS node with latest kernel DTS. This node has
two levels structure. The top level is used for hsiomix GPR register
access for USB wakeup which is not needed in u-boot. So we just
add a MISC driver to pass through to next DWC node to match the
XHCI driver

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>