configs: Layerscape: Remove the 'fdt_addr' env
On Layerscape platforms, the DTB is loaded from boot filesystem,
per the fdt_addr description in doc/README.distro, it must be
removed.

And on many platforms, like ls1046a, ls1088a, ls2088a and lx216xa,
the 'fdt_addr' pointed address is not accessible.
And with the current EFI boot process, since the EFI_LOADER and
CMD_BOOTEFI_BOOTMGR are enabled by default, if the EFI boot
components are not deployed in the boot filesystem, it will try
to get DTB at address 'fdt_addr' and then result in "SError" or
"Synchronous Abort":

Error log on ls1046ardb:
	=> run distro_bootcmd
	switch to partitions #0, OK
	mmc0 is current device
	Scanning mmc 0:1...
	libfdt fdt_check_header(): FDT_ERR_BADMAGIC
	Scanning disk esdhc@1560000.blk...
	Found 5 disks
	No EFI system partition
	"Error" handler, esr 0xbf000000
	elr: 00000000820704f4 lr : 00000000820080d4 (reloc)
	elr: 00000000fbd914f4 lr : 00000000fbd290d4
	x0 : 0000000064f00000 x1 : 00000000edfe0dd0
	x2 : 0000000000000000 x3 : 0000000000000000
	x4 : 00000000fbc2ee1a x5 : 000000000000000f
	x6 : 0000000000000000 x7 : 0000000000000008
	x8 : 0000000000000010 x9 : 0000000000000008
	x10: 0000000000000044 x11: 00000008ffff0220
	x12: 00000000fbdaa748 x13: 00000000fbda70f8
	x14: 00000000fbd21d20 x15: 00000000fbc194e8
	x16: 00000000fbd70fc8 x17: 0000000000000000
	x18: 00000000fbc1cdb0 x19: 00000000fbd21bf0
	x20: 0000000064f00000 x21: 00000000fbda6fb8
	x22: 0000000000000018 x23: 0000000000000018
	x24: 00000000fbde6344 x25: 0000000000000000
	x26: 0000000000000000 x27: 0000000000000000
	x28: 00000000fbc53660 x29: 00000000fbc19220

	Code: 7a419060 1a9f3000 a8c17bfd d65f03c0 (12800100)

	Resetting CPU ...

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>