MA-19331 imx8m: skip JR probe for android in uboot
As JR0 is reserved to secure world for HAB (secure boot),
commit 'ed618ce6a9 (LFU-219)' switch to use JR1 in uboot.
But this will cause conflict with Android as JR1 is
assigned to secure world for Trusty OS.

Since android doesn't use CAAM in uboot so we just skip the
JR probe in uboot to avoid conflicts.

Test: Boots on imx8mp.

Signed-off-by: Ji Luo <ji.luo@nxp.com>
Reviewed-by: Gaurav Jain <gaurav.jain@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>