LF-4677-4 spi: nxp_fspi: enable octal dtr support and fix odd address issue
Enable the octal dtr operation in driver and also fix the odd address
access issue, which happened in octal dtr mode only.

When FSPI running in octal ddr mode, it won't access the odd start
address A, but to access the previous 2 byte aligned start address A-1,
when access the nor chip via IPS. The code change fix the issue and
tested with mt35xu512aba, which is 128K erase size and easy to trigger
the issue when running UBIFS.

The code change can fix the read on odd address issue but it's still a
workaround. There are some cases can NOT be fixed, theoretically. For
instance, read CR register in OCTAL DTR mode with odd address. Since the
nor chip can only outputs the same bytes repeatedly when continuously
read, so there is no SW workaround, but we haven't seen any real cases
till now.

Signed-off-by: Han Xu <han.xu@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>