board: ls1088a: update ifc node name to be memory-controller
LF-4834: kernel boot failed

IFC-NOR and QSPI are muxed on SoC.
So disable IFC node in dts if QSPI is enabled or disable QSPI node in dts
in case QSPI is not enabled.
The nor-flash offset in uboot is invalid because of the following changes:
project: linux-nxp
* commit 'b6bd3c1f6ce17353601187c51424c18257735cc3':
  arm64: dts: freescale: update ifc node name to be memory-controller

Need to modify "ifc/nor" to "memory-controller/nor" in fdt_path_offset().

Signed-off-by: Jianpeng Bu <jianpeng.bu@nxp.com>