LFU-322 mmc: fsl_esdh_imx: Fix compatible issue in SD card init clock
In commit 6c9d901 (LFU-278-15 mmc: fsl_esdhc_imx: Send 80 clocks before
IDLE command), INITA was set to send init clock to SD card for power up
sequence. However, we find compatible issues on 8MQ EVK and 8QXP MEK with
some brands of SD cards.

The root cause is related with power up time.  According to spec, after
power stable, host shall supply at least 74 SD clocks to the SD card with the
maximum of 1ms. However, the power ram up time is related with the characteristic
of SD card. At the moment of sending 74 SD clocks, the power probably not ram up
to the operating level on the problematic cards. Then cause the cards not ready.

This patch changes to send SD clock with 1ms duration to replace 80 SD clocks
(0.2ms at 400Khz clock).
This way meets the spec requirement as well, and adds the margin for power ram up
time to be compatible with the problematic SD cards.  This is also aligned with
previous 2021.04 implementation which has FORCE clock always on.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-and-Tested-by: Haibo Chen <haibo.chen@nxp.com>