MA-20149 set fs type of android partitions
The usedata partition fs type is switched to be f2fs. Set it uboot for
the return result of "fastboot getvar" command, and the "fastboot format
userdata" can automatically format the userdata partition as f2fs.

For read-only partitions with android rootfs images, the fs type is
switched to be erofs. Althrough these partitions are dynamic partitions
which cannot be recognized by uboot, set their types still.

Change-Id: I393c18fc9f707f9a5d800e3721935a1624c0e0e9
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>