LFU-337 imx9: Remove DDRMIX slice power cycle
When A55 boot, DDRMIX is default power on. It is not needed to do
additional power off/on cycle to DDRMIX SRC slice.
This power cycle was thought to resolve DDR PHY accessing issue,
but actually the issue is caused by wrong DDR PLL setting which has
been fixed. So remove the unnecessary power cycle.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>