LF-6692-4 imx8ulp_evk: Disable the PMC thermal driver
On DBD_EN fused part, the upower space is defined to secure access
by TRDC. It is not good to open nonsecure access as the TRDC
granularity is 64KB, so that whole upower peripheral space will be
exposed.
We disable the PMC thermal driver, and will switch to use SCMI sensor
driver for thermal. This finally traps to ATF which calls the upower
API to get the temperature.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>