LF-7673-1 imx93: enable auto cmd12 for usdhc
On imx93, if config eMMC in HS400ES mode at 200MHz, after cmd18, send
cmd12 always meet command index error and command crc error. If downgrade
the clock rate to 100MHz, no this issue. And also no this issue if
emmc work in HS400 mode in 200MHz, so this should be related with the
usdhc data strobe logic. The root cause of this issue is still unknown,
IP team still in checking.

To workaround this issue, enable the auto cmd12 on imx93, this can pass
our stress test.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>