LFU-444-1: imx8mm(n)-evk: Add PHY GPIO reset timing
The base board of i.MX8MM/MN EVK has been re-designed due to EOL
component for the ethernet PHY. And the new PHY is RTL8211.
According to the datasheet RTL8211, it must be asserted low for
at least 10ms and at least 72ms "for internal circuits settling
time" before accessing the PHY registers. Otherwise the PHY does
not work correctly. The newly added properties do not affect the
AR8031 PHY, because these are generic properties and have no effect
on the reset of GPIO.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Ye Li <ye.li@nxp.com>