driver: net: ldpaa_eth: prevent dereference of NULL pointer
The ldpaa_eth_get_phy_mode_str() function could return a NULL pointer.
In this case, we would be dereferencing a NULL pointer.
Prevent this dereference from happening and set the PHY_INTERFACE_MODE_NONE as the
phy_mode in case we git this case.

This patch will not be needed once the following upstream patch will get
into our downstream branches.

123ca114e07e ("net: introduce helpers to get PHY interface mode from a device/ofnode")

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>