LFU-460 DTS: imx93-11x11-evk: add USDHC1 pinctrl setting in SPL stage
Android need to init the eMMC in SPL stage, when eMMC work at DDR52
in 8 bus width, always meet data transfer error, but when use 4 bus
width, issue gone. This is because in SPL stage, lack the USDHC1 pad
pinctrl config, so depends on the ROM config, and ROM only use 4 bus
width, so when try 8 bus width, meet issue.

This patch add the USDHC1 pad pinctrl setting in SPL stage, and has
no impact for the case which do not need to init the eMMC in SPL stage.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Acked-by: Ye Li <ye.li@nxp.com>