LFU-513 imx8: Fix the fuse used by secondary container offset
In SPL image-container.c, the incorrect fuse bits are used to
determine which offset to use when booting a secondary container.
This is generally not a problem if the fuse holding the offset is
all zeros, which is the case most of the time. If this is not the
case, issues can happen.

Signed-off-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Sebastien Haezebrouck <sebastien.haezebrouck@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>