LFU-552-3 misc: ele_api: Update ELE read common fuse API
On iMX8ULP, the word index 1 is used to read OTP_UNIQ_ID with 4 words
data responsed. However this special index does not for OTP_UNIQ_ID
on iMX93, it is just an normal fuse word index. So update the check
to avoid problem when reading from fuse word 1 on iMX93.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>