MA-21906-4 imx95: support Trusty OS
Use 'elemu1' in spl stage to sync the MU used in Trusty
OS, so we can derive the same HUK in spl and in Trusty.

And because the TEE was moved to secure memory of DDR
(0x8C000000 - 0x8DFFFFFF) which is not in the scope of
the DRAM space in u-boot, we should skip removing TEE
memory region in such case.

Signed-off-by: Ji Luo <ji.luo@nxp.com>