LFU-826 cmd_qb: Add optional args to specify where to save qb data
The qb save command saves the data to the boot device, but the
user can also specify a different location, which means that this
command can now also be used from within a uuu script. The same
also goes for qb erase.

Usage: qb save/erase [interface] [dev].
For example, to save the qb data to eMMC: qb save mmc 0

Signed-off-by: Simona Toaca <simona.toaca@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Tested-by: Viorel Suman <viorel.suman@nxp.com>